requirements.txt
einops==0.8.1
importlib-metadata==8.0.0
jaraco.collections==5.1.0
librosa==0.11.0
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.8.0
torchaudio==2.8.0
pytorch-lightning
laion-clap
pedalboard
numpy
pyyaml
tensorboard